requirements/test.txt
# Test dependencies go here.py
-r base.txt



coverage
flake8 # pyup: != 2.6.0
django-test-plus
factory-boy

# pytest
pytest-django
pytest-sugarp
